requirements.txt
fastapi
uvicorn
python-multipart
pymongo
cloudinary
python-dotenv
opencv-python
numpy
scikit-image
pytesseract
scipy